requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

python-telegram-bot==21.9
yt-dlp==2024.12.13

httpx==0.28.0
anyio==4.7.0